refactor(View): Color를 도메인에서 뷰로 이동
